6_joints_3_force_comp_1_force_vec_sum
Another alteration to the data.
Now we have all 6 joints, why? because I think the previous data may be corrupted. 
Explain- When collecting robot position and force data kinaesthetically, There were joint deviations, particularly on joint 4-the second to last joint before the end-effector. 
It is thought that any joint deviation as a result of forces tugging on the robot will provide extra information to the model which allows it to learn easier, but this information won't be available when planning future robot trajectories. 
Unplanned joint deviations are just that- unplanned; therefore, having a machine learning model that is potentially using this joint deviation as a key feature will not work in deployment. 
In this dataset, I include all 6 joints, any of which can be dropped by the model code. 
previously, the model has only used joints 0,2,4 and 5, as joints 1 and 3 were highly correlated with joint 2, but removing joint 4 would only leave joints 0, 2 and 5.
It's possible that joint 5 may also be corrupt but I suspect this corruption will mainly be in torque space, which we are not as concerned with. At least, not yet.